cody: deduplicate summariseAttribution implementations (#52197)
We don't expect to keep sharing this implementation in the future, but
for now it makes sense to atleast deduplicate. This was motivated by me
changing the shape of the API and being annoyed by the duplication. I
didn't finish changing the shape, but in a future PR will be adding more
fields to the newly introduced Attribution interface.

Test Plan: CI