Make cody requires a verified email consistent with site resolver (#52217)
Currently the `useIsCodyEnabled` hook and the server side checks for if
the site requires the user to have a verified email are inconsistent.

This adds to the JSContext a property that indicates if the site
requires verified emails and then uses that instead of directly checking
against sourcegraph.com. This consistently helps App because it usually
passes cody requests though to dotcom.