embeddings: support delta updates (#52158)
With this change we only embed new or modified files.

I took inspiration from searcher's hybrid search. Specifically we reuse
the logic to call git diff and parse the output.

Approach:
- call git diff on new vs. old commit and get modified, added, deleted
files
- index modified and added files -> "delta index"
- download existing index
  - remove modified and deleted files
  - append delta index
- upload index

Things to consider:
- the embeddings are gob-encoded and gob doesn't support appends
- the store doesn't support appends either
- We could just upload the delta, but that would mean we have to think
about compaction and tombstones.
- Even though we have an additional download of the embeddings from the
store, this should be more than compensated because we don't send the
entire code to the embeddings provider.

Most of the code is behind a feature-flag.