App: add updater endpoint for Sourcegraph Updates (#51651)
Moved app update endpoint to be next to the usual `.update` endpoint,
with resolvers that can be swapped out. Static resolver for local
testing and GCS on dotCom to get the update manifest. If one or more
resolvers fails, we use a ` NoopUpdateChecker` which tells all clients
that there are no updates - until we fix whatever we're experiencing

## Test plan
Tested locally
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->