[fix] switch id column of user_repo_permissions table to bigint
This is to avoid overflow on the primary key, which defaults to a sequence
value. We are using insert queries with ON CONFLICT DO SOMETHING. These types
of queries first try to insert the row, hence they compute the next value
in sequence. Even if we do not actually insert any rows because of conflict.

With user_repo_permissions table this might become a problem, as can be seen
in [#inc-209-fiverr-user-permission-sync-is-broken](https://sourcegraph.slack.com/archives/C058XNX78D8/p1684785371774189)