llm-proxy: only consume rate limit on successful upstream request (#52238)
Closes https://github.com/sourcegraph/sourcegraph/issues/51879. A
"commit" callback is now returned from `TryAcquire` so that unless the
end response to the client is a healthy response, we do not consume the
rate limit. We still check if rate limit has been exceeded in the
hot-path of `TryAcquire`. This can cause some overallowance, but we're
okay with that.

## Test plan

Unit tests