Cody completions: Over-fetch manual completions but hide empty results (#52304)
Something really annoying with manually triggered completion is that it
would include empty results (sometimes the response from the LLM are not
great). This is fine for inline completions since those will be filtered
out automatically but for manual ones, it would show an empty item in
the sidebar which looked odd.

To smooth over some of these edges, this over-fetches the number of
requested completions by two (ONLY for manual completions), filters out
empty results, and caps the number to three (so it feels more reliable).

I think in the future we might want to rather stream those in and start
additional requests only when some requests come back empty. This
feature is not used much yet though so this seems reasonable to me.

## Test plan

<img width="2029" alt="Screenshot 2023-05-23 at 14 56 14"
src="https://github.com/sourcegraph/sourcegraph/assets/458591/b006ee7a-1a02-4261-9245-dd8c5cdfad26">

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->