guardrails: GraphQL API for attribution
This is an initial port of guardrails attribution to the Sourcegraph
backend. It still has quite a bit of follow-up work, but this
establishes a base working version and structure.

The API introduced is marked as experimental and is intended to be used
directly from Cody clients. Additionally the introduced services are
intended to be used by our Chat and Completions endpoints for filtering.

Test Plan: Right now I have just tested it compiles. I have high
confidence from CI this breaks nothing else, and using the experimental
nature to write tests later as I get more confident in the design.