app: show helpful error messages in cody window if cody can't be used yet (#52344)
The Cody window has three states where it can't be used:
1. Setup not complete
2. Sourcegraph.com account not connected
3. Sourcegraph.com account does not have a verified email

This PR adds helpful error messages to get the user to a good state to
use the standalone Cody window:
1. When setup is not complete, the Cody window will display a button to
open setup; clicking the button will focus on the main window where the
setup should be running. Once the setup is complete, the Cody window
automatically detects this and moves on from this state.


https://github.com/sourcegraph/sourcegraph/assets/206864/67402029-8677-48fb-8b45-d5639a916ec0


2. When a Sourcegraph.com account is not connected, the Cody window will
display a button to connect; this opens Sourcegraph.com in the browser
with the auth flow. Once the auth flow completes, the main app window
makes a call to Tauri to refresh the Cody window so it gets the new
config state.


https://github.com/sourcegraph/sourcegraph/assets/206864/0047c927-360b-48c4-b848-b1122e9ed3a6

3. When the connected Sourcegraph.com account does not have a verified
email, the Cody window will display a button to go to verify the email.
Unfortunately, we can't automatically refresh when this is resolved
since this is done external to the app, so we tell users to restart the
app when complete.


https://github.com/sourcegraph/sourcegraph/assets/206864/800e7ae5-a247-431b-973e-53b8ee2b1651

Additionally, the following polish fixes were made:
- The Cody window no longer displays the setup experience when setup has
not been completed
- The unverified email banner is not displayed in App because the way
the link is rendered does not work in App

## Test plan

See videos above