Cody: Add tests for Inline Assist (#52129)
Close https://github.com/sourcegraph/sourcegraph/issues/51940

This PR adds a test for the lineTracker function used by `Inline
Assist`. It is a major component that tracks changes made in the editor
where Cody is working on to make sure the correct lines of code are
getting highlighted, decorated, and replaced.

#### ~SOS~

This is the first time I work on adding a new test for function that has
dependencies, and I can't seem to get it to work when importing the
`vscode` package and the lineTracker function. This is why I duplicated
the origin function in the test file instead of using import. Any advice
on how to resolve this is greatly appreciated 🙇

Update: Resolved

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

The newly added tests have all passed.

Here is another PR for adding e2e test for Inline Assist:
https://github.com/sourcegraph/sourcegraph/pull/52337