Merge extsvc Kind and Type enums (#52211)
External Services come in many flavors, from GitHub repositories to Rust
packages. To support this wide variety of services, schemas and structs
are created to define the capabilities and properties of each service.
Different strings are used to define what kind of service each one is.
There is a string in GraphQL and the configuration files that is labeled
as the "Kind" of service is it, and there is a string in the database
that is labeled as the "Type" of service it is. There is a lot of
converting between those two, and one has to make sure to use the
correct label in the correct situation.

OK, I'm going to shut up now and let Cody describe the situation.

Cody, why are there separate "Type" and "Kind" variables defined? What
is the difference?

> The `Type` and `Kind` variables in the extsvc package represent the
same external service concepts, but are used in different contexts:
> 
> - `Kind` variables are used in the Sourcegraph API and configuration.
For example, the kind field in external service configuration specifies
the kind of external service.
> 
> - `Type` variables are used in the Sourcegraph database schema. For
example, the `external_service_type` column in the `repo` table stores
the type of an external repository.
> 
> So in short:
> 
> - `Kind` is used in the API and config.
> - `Type` is used in the database schema.
> They represent the same concepts, just used in different contexts.
> 
> The reason for having separate `Type` and `Kind` variables is mostly
historical. Going forward, the `extsvc` package is moving to using
`ServiceVariant` enums instead, which unify the Type and Kind concepts.

Yep, going forward we'll be able to clean this mess up by consolidating
on a `ServiceVariant` type that describes the type/kind of service.

This PR adds the `ServiceVariant` type, with supporting enum and a map
to hold the values associated with the variant.

The rest of the code continues to use the `Type` and `Kind` variables,
but those variables are now backed by `ServiceVariant`. Conversion will
happen later, probably in stages, because there are hundreds of usages
across dozens of files.

## Test plan

All existing tests should pass

---------

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>