App - Ensure verify email check always returns true or false  (#52425)
App checks if the current user has a verified email address for sg.com.
If the token is no longer valid this check fails, and causes the the JS
Context to be built without a user object. This puts the app into a not
signed in state and the user can't resolve the issue.

This change ensures that the check against sg.com will always return
without an error and defaults to false if an error occurs. While this
may result in the user seeing message that they need to verify their
email at least the user can correct the issue.

A follow up should be to allow the user to re-connect to dotcom if their
token is no longer valid.