Cody completions: Log multi-line completion requests and results (#52449)
Adds a flag to the event logs so we know how often multi-line
suggestions are triggered.

Logging how often they are accepted is a bit tricker and requires some
resturcturing. For now we do not know for an accepted event wether it
was a multi line or a single line completion. Will fix it up later (but
I plan to cut a release in a bit)

## Test plan

<img width="1092" alt="Screenshot 2023-05-25 at 14 14 15"
src="https://github.com/sourcegraph/sourcegraph/assets/458591/da3ed4c4-b0c2-4906-9680-b6324c483c04">


<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->