[fix] use same where condition in user analytics as in pings for mau (#52306)
Previously we used a different condition to calculate the MAU and other
stats in user analytics page. Now we use the same condition as we use
for pings. We also calculate everything using UTC now, so both should
give the same value.

## Test plan

Tested locally and also with data from Tinder. The new query for MAU
seems to be returning the correct data for us, I manually
cross-referenced the old query vs the new. Apart from the timezone
difference, the new query is more precise as in it filters the
sourcegraph-operator users and non-user events in the same way as we do
for the pings.

(cherry picked from commit 8d04aa1e3f851f01c68525470473f0ca546dfb3e)