Tracing: add `trace.FinishWithErr` (#52478)
This just adds a little helper method for a very common operation when
using tracing: finish the span and set its error.

Basically, this turns a 4-liner into a 1-liner. This is nice because
tracing can add quite a bit of noise to a function. Also, it's directly
deferrable so it's less error-prone.