gitserver/server: Move runWith and other related funcs and types to common package (#52334)
## Why

As part of #40330 I need to use `runWith` and `GitCommandError` in the
`cmd/gitserver/server/perforce` package which will be introducded in
#51860. Since it is declared in the `cmd/gitserver/server` package,
moving it to `cmd/gitserver/server/common` helps avoid a
cyclic-dependency issue. The large diff is because I had to move all the
related symbols that `runWith` depends on.

Follow up to #52293 as I missed this requirement earlier with all the
moving pieces.

👉 Also, I did some code reorganisation to have the public API on the top
of the `run.go` file.

Also skipping a lint check for the `internal/trace/ot` package because I
do not want to change behaviour in this PR and want to restrict this to
only a code reorganisation.