[fix] add back the condition to filter out backend events
This is for usage analytics in the product. Previously we were filtering
out the backend events, but as part of #52306 this was removed, because
I believed it shold not be there, as it was not part of the BuildCommonUsageConds
in the event_logs. However event_logs also add this condition directly
in the query string, which I did not notice.