gitserver: Use a generic FIFO queue for cloning repos (#52574)
Part of #40330

## Why

In #51860 we are adding a new queue to perform mapping of perforce
depots. The queue implementation is a copy-paste of the existing
`cloneQueue`. Instead of copy-pasting, we take a short detour to
implement the queue with generics so that it may be used in the new
feature as well.

No functional change to how repos are cloned is being introduced here.