gitserver: use cat-file for reading blobs containing ..
We special case ".." in path to running a less efficient two commands.
For other paths we can rely on the faster git show.

git show will try and resolve revisions on anything containing "..".
Depending on what branches/files exist, this can lead to:

- error: object $SHA is a tree, not a commit
- fatal: Invalid symmetric difference expression $SHA:$name
- outputting a diff instead of the file

The last point is a security issue for repositories with sub-repo
permissions since the diff will not be filtered.

Test Plan: wrote tests and test e2e