[fix] fixes for mau calculations in product analytics (#52581)
Fixes MAU calculation in usage analytics in the product. 

Previously we were filtering out the backend events, but as part of
#52306 this was removed, because I believed it shold not be there, as it
was not part of the BuildCommonUsageConds in the event_logs. However
event_logs also add this condition directly in the query string, which I
did not notice.

Also, our timestamp calculation suffered from an edge case, where the
from time was not calculated right, e.g. on dates like
`2023-04-30T23:59:00Z`.

## Test plan

Tested locally and with customer data + added unit tests.