Search: fix UseKeywordScoring hack (#52649)
We recently enabled Zoekt's new BM25 scoring for the `keyword` search
type. We enabled the option using feature flags, which is hacky because
users will never be touching this setting.

This PR refactors all Zoekt-related jobs to use the
`search.ZoektParameters` struct. This lets us pass the flag to Zoekt
directly when constructing jobs.

Follow-up to #52233