app: fix asset resolution for builds outside of bazel (#51549)
We recently changed how assets are resolved in the backend. If the
backend is built with Bazel it just works, as well as when you run in
Dev mode.

But if the backend is built the normal go way, the assets are not where
the backend expects it to be. This fixes the frontend code to resolve
the assets directory depending on the environment
## Test plan
Check assets gets rendered with the following commands
* `sg start`
* `bazel build //enterprise/cmd/sourcegraph:sourcegraph`
* `./enterprise/dev/app/build-release.sh`
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->