ci: propagate exec errors in the report-bundle-diff script (#52740)
`shelljs.exec` returns the result with the error code but does not
throw. This PR replaces it with `child_process.execSync`, which throws.
This is what we expect when we iterate over `revisions` with bundle size
stats on `main`. Should fix the issue reported in [this Slack
thread](https://sourcegraph.slack.com/archives/C04MYFW01NV/p1685575081237319).