server: do not use RunWith for parsing git log
RunWith is supposed to be used for remote commands like git fetch. You
should follow up and remove it from common, it is an internal detail of
server.

The root cause of the issue was RunWith writes to progress writer in
another goroutine, but you would close the channel once RunWith
returned.

This commit does two things in the process of fixing this. It simplifies
code by just directly running the scanner on command output. This lets
us avoid all the code around goroutines and channels. Additionally there
is another bug since it was reading line by line which assumes the
commit messages are always just one line. It is pretty easy to just
adapt ScanLines to look for nulls.

Test Plan: "go test -race -count=50 ./cmd/gitserver/server/perforce"