productsubscription: move token generation to licensing, change token prefix (#52811)
With #52634, we are making the product subscription access token
explicitly the license key, rather than hiding it as an implementation
detail. This change moves access token generation to the `licensing`
package so we can generate it from anywhere, and creates a new token
prefix `slk_` ("S"ourcegraph "L"icense "K"ey) to make the relationship
explicit. We retain support for the old prefix for now.

## Test plan

Tests. Added a new test to cover back-compat with the old format as
well.