cmd/gitserver: Move common.RunWith inside server
In this commit we move common.RunWith inside the server package and
rename them to make them private.

In #52334, we moved this to the common package because it was required
in #51860. However after iterating further we do not need it any longer
we move this back into server along with all the other related symbols.