search: return error instead of panic in NewFileContainsFilterJob
I'm not too familiar with our job code, but I believe we expect all
patterns in them to already be valid regexes. However, we are getting
panics in production. It is likely better to just fix the root cause
here since panic is caught by the http handler. But errors feel better.

The example panic I saw in prod had this message:

  handler panic: regexp: Compile(`(?i:r.Get("/")`): error parsing regexp: missing closing ): `(?i:r.Get("/")`

Test Plan: go test