Bag panic root cause fix. (#52877)
Part of #52911 

The bug is when user references augmenting errors. Previously we already
linked `resolvedUserID` to the reference (like a handle or email) but
actually getting all the other references known to sourcegraph fails,
and we have effectively a dangling reference that manifests as nil
pointer during map lookup.

## Test plan

TODO: Follow up with unit tests that try to expose issues that could
cause panics and errors.