ci: fix image scanning dependencies
Adds two explicit step dependency for symbols and syntax-highlighter.

Previous code for handling dependencies is a bit fragile, which leads to
corner case where it tried to scan an image that wasn't built, which is
confusing to read.

This will be cleaned up in the future by https://github.com/sourcegraph/sourcegraph/issues/52833

Test Plan: previewed locally + main-dry-run