nix: develop uses non static binaries
In our flake we expose a bunch of statically compiled binaries. We
introduced this to help us ship those binaries that are quite hard to
build outside of nix. With our recent re-organization we started using
them. The downside of this is we now have to do things like recompile
ocaml. Even on my beast linux desktop this feels like overkill.

This commit switches us to using the normal nix packages. Additionally
it:
- removes p4-fusion from our devenv since no nix users dev with it.
- re-introduces the universal-ctags wrapper

Test Plan: checked ctags in the devenv:

  nix develop
  file $CTAGS_COMMAND
  which universal-ctags