bazel: update gazelle repos
I tried this out because I was getting some test failures in go test
that looked like a dependency issue and was wondering if CI wasn't
picking it up. Lets see if CI is green.

Additionally I updated outdated references to how we used to run
gazelle.

Test Plan: CI