Cody backend: unify context fetching behind a single endpoint (#52632)
The primary goal of this PR is provide an easy fallback to keyword
search for repos that don't have embeddings enabled. To enable this, I
created a new, unified context fetching API so that we can hide the
details of how we populate context from the client (they don't need to
care whether a repo has embeddings or not).