search: remove repeated arguments in client.New
This commit updates all calls to client.NewSearchClient to not specify
zoektStreamer, searcherURLs and searcherGRPCConnectionCache anymore. We
always passed exactly the same values everytime, so we might as well
simplify the call site.

In tests we only ever mocked zoekt, so we add a new constructor for
tests which capture the pattern used there.

Additionally we rename the method from client.NewSearchClient to just
client.New to avoid the stutter. Call sites generally make it clear the
variable being captured is related to search, so this reads better.

Test Plan: CI